Create BTC low Fee_V1
This Python code is implementing a "TitleChain" class to anchor arbitrary data to the Bitcoin blockchain by creating Bitcoin transactions with the data embedded.

Here is what it is doing:

    Imports necessary libraries like bitcoinlib for wallet and transaction creation, and esplora for fee estimation
    The TitleChain class initializes with a wallet name and network
    estimate_optimal_fee_rate() uses Esplora API to get the latest BTC network fee rate, and adjusts it lower for cost optimization
    anchor_to_bitcoin() creates a Bitcoin transaction with SegWit outputs, adds the data using nulldata outputs, signs and broadcasts the transaction
    The transaction ID serves as an anchor proving the data was embedded on-chain at a point in time

To use it:

    Create a TitleChain instance with a wallet name
    Call anchor_to_bitcoin() with a list of data entries to anchor
    The data will be anchored on-chain and the transaction ID returned
It allows efficient anchoring of arbitrary data on the Bitcoin blockchain by embedding it directly in transactions. Could be used for timestamping, proof of ownership, document authenticity etc.